ref(mode): optimize service catalog lookup construction
